fix: wrap url state in suspense
